requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
idna==2.10
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
